+++ test/Feature/CleanRegex/_prepared/inject/pcre/PatternTest.php
@@ -21,7 +21,7 @@ class PatternTest extends TestCase
     public function shouldUsePlaceholder(string $pattern, string $expected)
     {
         // when
-        $pattern = Pattern::builder()->pcre()->inject($pattern, ['X']);
+        $pattern = Pattern::pcre()->inject($pattern, ['X']);
 
         // then
         $this->assertSamePattern($expected, $pattern);
@@ -45,7 +45,7 @@ class PatternTest extends TestCase
     public function shouldNotMistakeLiteralForPlaceholder(string $pattern, string $expected)
     {
         // when
-        $pattern = Pattern::builder()->pcre()->inject($pattern, []);
+        $pattern = Pattern::pcre()->inject($pattern, []);
 
         // then
         $this->assertSamePattern($expected, $pattern);
@@ -68,7 +68,7 @@ class PatternTest extends TestCase
     public function shouldNotMistakePlaceholderInCommentInExtendedMode()
     {
         // when
-        $pattern = Pattern::builder()->pcre()->inject("%You/her #@\n her?%x", []);
+        $pattern = Pattern::pcre()->inject("%You/her #@\n her?%x", []);
 
         // then
         $this->assertSamePattern("%You/her #@\n her?%x", $pattern);
@@ -80,7 +80,7 @@ class PatternTest extends TestCase
     public function shouldUsePlaceholderInCommentInExtendedMode_butExtendedModeIsSwitchedOff()
     {
         // when
-        $pattern = Pattern::builder()->pcre()->inject("%You/her (?-x:#@\n) her?%x", ['X']);
+        $pattern = Pattern::pcre()->inject("%You/her (?-x:#@\n) her?%x", ['X']);
 
         // then
         $this->assertSamePattern("%You/her (?-x:#X\n) her?%x", $pattern);